Changed how items work in an inventory.
Items now tick when in the players' inventory or when equipped. Some
items may do things only when equipped, only when not equipped, or can
ignore equipped completely. Changed how the inventory is laid out so the
controls will be easier to use when navigating. Changed how lighting
works when it comes to light sources in tiles that aren't affected by
lighting (see torch wall), tiles now have givesLight and takesLight
instead of a single lightable function.