Added combat mechanics.
Finished adding a skeleton for combat testing purposes. The test sword
has been renamed to badsword and temporarily deals acidic damage instead
of sharp damage to test skeleton deaths (0.2x damage for sharp on
skellies). Modified how the tex variable in the necklace/sword is
initialised in preperation of randomised sprites to prevent remembering
which sprite goes to which item. game.jar not updated because current
version crashes on skelly death.