Pickaxe, level gen, and void improvements
Void 'tiles' are now rendered as if they are the tile specified as a
map's base tile with no light (the actual void remains black to mark a
level edge, looks really nice).
Pickaxe code no longer references any specific tile except for TileVoid,
which will be the "universal constant" for a lack of a tile.
Tiles now have a function to state if something is allowed to break a
tile next to it (good for doors).
The level generator now has a 1 in 3 chance of letting a hallway going
through another hallway that it just encountered during generation.